+++ Changelog.md
@@ -18,6 +18,7 @@
 * Add support of parsing more text directions in cell
 * Add support of `show_category_name` and `show_series_name` to chart label properties
 * Add method to 3 base formats to check if it contain any user data `#with_data?`
+* Correct parsing of paragraph properties and run properties
 
 ### Fixes
 * Fix parsing document style id - it can be string, not only digit